adding update functions

+++ src/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: malourei <malourei@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/15 00:32:30 by malourei          #+#    #+#             */
-/*   Updated: 2024/11/27 20:16:01 by malourei         ###   ########.fr       */
+/*   Updated: 2024/11/28 21:21:35 by malourei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ void	*ft_calloc(size_t nmemb, size_t size);
 char	*ft_strcpy(char *dst, const char *src);
 char	*ft_strcat(char *dest, const char *src);
 char	*ft_strrchr(char *s, char c);
-int		ft_strlen(char *str);
+size_t		ft_strlen(char *str);
 int		ft_strcmp(char *c, char *d);
 char	*ft_strdup(char *s);
 char	*ft_strncpy(char *dest, const char *src, size_t n);
